Pivot wide format ausTrait data into a long format

trait_pivot_longer "gathers" wide format data into a "tidy" format This function converts the data into long format where observations are on different rows and the type of observation is denoted by trait name. In other words, trait_pivot_longer reverts the actions of trait_pivot_wider

trait_pivot_longer(data)

Arguments

data
A tibble in wide format generated from trait_pivot_wider - see example

Value

A tibble in long format

Author

Daniel Falster - daniel.falster@unsw.edu.au

Examples

if (FALSE) {
data <- austraits$traits %>% 
filter(dataset_id == "Falster_2003")
data #long format 
traits_wide <- trait_pivot_wider(data) 
traits_wide #wide format

values_long <- trait_pivot_longer(traits_wide)
}